# backend/openapi/qlix-developer-v1.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Qlix Developer API",
    "version": "1.0.0",
...
  },
  "servers": [
    { "url": "/api/v1", "description": "Versioned API root" }
  ],
...
  "components": {
    "securitySchemes": {
      "ApiKeyBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "qlix_live_*",
        "description": "API key issued from the Qlix console"
      },
      "ApiKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Api-Key"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "object",
            "required": ["code", "message"],
            "properties": {
              "code": { "type": "string", "example": "insufficient_scope" },
              "message": { "type": "string" }
            }
          }
        }
      },
      "Agent": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "userId": { "type": "string" },
          "orgId": { "type": "string", "nullable": true },
          "did": { "type": "string" },
          "publicKey": { "type": "string" },
          "name": { "type": "string" },
          "description": { "type": "string", "nullable": true },
          "status": { "type": "string" },
          "runtime": { "type": "string", "enum": ["cloud", "local", "hybrid"] },
          "model": { "type": "string" },
          "localInferenceMode": { "type": "string", "nullable": true, "enum": ["local_llm", "cloud_api"] },
          "llmMode": { "type": "string", "enum": ["direct", "proxy"] },
          "llmProvider": { "type": "string", "enum": ["exora", "openrouter"] },
          "permissionScopes": { "type": "array", "items": { "type": "string" } },
          "jitScopes": { "type": "array", "items": { "type": "string" } },
          "alwaysScopes": { "type": "array", "items": { "type": "string" } },
          "agentKind": { "type": "string", "enum": ["standard", "org_brain"] },
          "toolProfile": { "type": "string", "enum": ["minimal", "coding", "lean", "full"] },
          "createdAt": { "type": "string", "format": "date-time" },
          "lastActive": { "type": "string", "nullable": true }
        }
      },
...
      "ConfirmNameRequest": {
        "type": "object",
        "required": ["confirmName"],
        "properties": {
          "confirmName": { "type": "string" }
        }
      },
...
      "JitDecideRequest": {
        "type": "object",
        "required": ["jitRequestId", "approved"],
        "properties": {
          "jitRequestId": { "type": "string", "format": "uuid" },
          "approved": { "type": "boolean" }
        }
      },
      "JitGrant": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "scope": { "type": "string" },
          "agentId": { "type": "string" },
          "expiresAt": { "type": "string", "format": "date-time", "nullable": true }
        }
      },
      "JitPending": {
        "type": "object",
        "properties": {
          "jitRequestId": { "type": "string", "format": "uuid" },
          "agentId": { "type": "string" },
          "scope": { "type": "string" },
          "scopeLabel": { "type": "string" },
          "context": { "type": "string" },
          "runId": { "type": "string", "nullable": true },
          "conversationId": { "type": "string", "nullable": true },
          "sourceChannel": {
            "type": "string",
            "description": "Where the run started: api, web, whatsapp, slack, telegram, local, cron"
          },
          "requestedAt": { "type": "string", "format": "date-time" }
        }
      },
      "UsageSummary": {
        "type": "object",
        "properties": {
          "billingCycle": { "type": "string", "example": "2026-08" },
          "summary": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "agentId": { "type": "string" },
                "agentName": { "type": "string" },
                "totalRuns": { "type": "integer" },
                "totalTokens": { "type": "integer" },
                "totalCostUsd": { "type": "string" }
              }
            }
          }
        }
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "Missing, invalid, or revoked API key / session",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
      "Forbidden": {
        "description": "Missing scope, role, or route not in the Developer API",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
...
    },
    "parameters": {
      "AgentId": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "string" }
      },
      "AgentIdAlt": {
        "name": "agentId",
        "in": "path",
        "required": true,
        "schema": { "type": "string" }
      },
...
      "GrantId": {
        "name": "grantId",
        "in": "path",
        "required": true,
        "schema": { "type": "string" }
      },
...
    }
  },
  "security": [{ "ApiKeyBearer": [] }, { "ApiKeyHeader": [] }],
  "paths": {
...
    "/agents/{id}": {
      "get": {
        "tags": ["Agents"],
        "summary": "Get agent detail",
        "operationId": "getAgent",
        "parameters": [{ "$ref": "#/components/parameters/AgentId" }],
        "responses": {
          "200": {
            "description": "Agent detail",
            "content": { "application/json": { "schema": { "type": "object", "properties": { "agent": { "$ref": "#/components/schemas/Agent" } } } } }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        },
        "x-qlix-scopes": ["agents:read"]
      },
      "delete": {
        "tags": ["Agents"],
        "summary": "Delete agent",
        "operationId": "deleteAgent",
        "parameters": [{ "$ref": "#/components/parameters/AgentId" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ConfirmNameRequest" }
            }
          }
        },
        "responses": {
          "204": { "description": "Deleted" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        },
        "x-qlix-scopes": ["agents:write"]
      }
    },
...
    "/jit/decide": {
      "post": {
        "tags": ["JIT"],
        "summary": "Approve or deny a JIT request",
        "operationId": "decideJit",
        "description": "Decide a pending JIT request created by a running agent. Use this when the run was started with this API key (or from the console). WhatsApp/Slack-originated runs are decided in those chats instead.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/JitDecideRequest" }
            }
          }
        },
        "responses": {
          "200": { "description": "Decision recorded" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        },
        "x-qlix-scopes": ["jit:decide"]
      }
    },
    "/jit/grants": {
      "get": {
        "tags": ["JIT"],
        "summary": "List active session grants",
        "operationId": "listJitGrants",
        "responses": {
          "200": {
            "description": "Grants",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "grants": { "type": "array", "items": { "$ref": "#/components/schemas/JitGrant" } }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        },
        "x-qlix-scopes": ["jit:decide"]
      }
    },
    "/jit/grants/{grantId}": {
      "delete": {
        "tags": ["JIT"],
        "summary": "Revoke a session grant",
        "operationId": "revokeJitGrant",
        "parameters": [{ "$ref": "#/components/parameters/GrantId" }],
        "responses": {
          "200": { "description": "Revoked" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        },
        "x-qlix-scopes": ["jit:decide"]
      }
    },
    "/jit/pending": {
      "get": {
        "tags": ["JIT"],
        "summary": "List pending JIT requests",
        "operationId": "listPendingJit",
        "description": "Lists pending JIT approvals for the key's user. Omit filters to see every waiting request. Pass `runId` to scope to one run, or `conversationId`+`agentId` for a chat thread. Approvals for API-started runs appear here and as `jit_approval_pending` on `GET /agents/{id}/runs/{runId}/stream`.",
        "parameters": [
          { "name": "runId", "in": "query", "schema": { "type": "string" } },
          { "name": "agentId", "in": "query", "schema": { "type": "string" } },
          { "name": "conversationId", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Pending requests",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "pending": { "type": "array", "items": { "$ref": "#/components/schemas/JitPending" } }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        },
        "x-qlix-scopes": ["jit:decide"]
      }
    },
...
    "/usage/summary": {
      "get": {
        "tags": ["Usage"],
        "summary": "Usage summary",
        "operationId": "usageSummary",
        "parameters": [
          { "name": "billingCycle", "in": "query", "schema": { "type": "string", "example": "2026-08" } }
        ],
        "responses": {
          "200": {
            "description": "Usage summary",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UsageSummary" } } }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        },
        "x-qlix-scopes": ["usage:read"]
      }
    },
    "/usage/agents/{agentId}": {
      "get": {
        "tags": ["Usage"],
        "summary": "Per-agent usage breakdown",
        "operationId": "usageForAgent",
        "parameters": [{ "$ref": "#/components/parameters/AgentIdAlt" }],
        "responses": {
          "200": { "description": "Per-run usage" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        },
        "x-qlix-scopes": ["usage:read"]
      }
    },
...
  }
}